Changes
The following  bug was  approved to be released on version 8.0 with 1 / 0:

 #880511217347727460: Obitos Muugen Tsukuyomi tends to bug out and give permanent vision over the area

tonton-bot
BOT
 — Today at 20:59
The following  bug was  approved to be released on version 8.0 with 2 / 0:

 #880200396243607562: if you play konan and kill someone in missions using W, you bug yourself
[21:01]
The following  bug was  approved to be released on version 8.0 with 5 / 0:

 #880189593859784704: some spells still bug rin go fix metal

React with  to increase the priority of this bug. Use !reply 880189593859784704 to reply to this bug.
==================================
Reply #1: Young Kabuto R
[21:06]
The following  bug was  approved to be released on version 8.0 with 2 / 0:

 #880126579609591848: fix Rock Lee gates auto attacks being able to trap himself in dark areas
[21:10]
The following  bug was  approved to be released on version 8.0 with 2 / 0:

 #879869033254617089: fix sage naruto rasengan flying and then stopping randomly, when
[21:10]
The following  bug was  approved to be released on version 8.0 with 1 / 0:

 #878921639096631336: Animal Path does not have Triple Kill line

React with  to increase the priority of this bug. Use !reply 878921639096631336 to reply to this bug.
==================================
[21:15]
The following  bug was  approved to be released on version 8.0 with 3 / 0:

 #878816923738046494: reflection dmg isn't being applied to dmg over 4k
[21:19]
The following  bug was  approved to be released on version 8.0 with 2 / 0:

 #880125536666550324: Rock Lee can't suicide with Kaguya in the S rank with his T cause then the S rank drops in the midlane and he dies and can go back into the S rank and do the same thing over again.
[21:19]
The following  bug was  approved to be released on version 8.0 with 4 / 0:

 #870427386993315911: Fix bug that allows for a team to get 2 S rank items

React with  to increase the priority of this bug. Use !reply 870427386993315911 to reply to this bug.
==================================
Reply #1: metal knows how to do it
[21:23]
The following  balance issue was  approved to be released on version 8.0 with 1 / 3:

 #880532217997193236: Kisames Mugensames sharks should deal dmg below him